+++ server/graphql/mutations/documentType/deleteDocumentType.js
@@ -11,7 +11,10 @@ const deleteDocumentType = {
       type: GraphQLString,
     },
   },
-  resolve: async (parent, args) => {
+  resolve: async (parent, args, req) => {
+    if (!req.isAuth) {
+      throw new Error('unAuthorized')
+    }
     const deleted = await DocumentTypeModel.findByIdAndDelete({ _id: args._id });
     if (!deleted) {
       throw new Error('Error');
